Move stabilization to interface.
